scripts/socket-security/requirements.txt
# main
requests
rich


# test
python-dotenv
pytest
